Add some checks when sending data, allocating blocks, and use size_t when needed.

+++ audio_packet.h
@@ -19,6 +19,6 @@
 
 struct server;
 
-int audio_received(char *in, int len, struct server *s);
+int audio_received(char *in, size_t len, struct server *s);
 
 #endif
